Searcj_proj Complete
Completed my small program to search for files with given extensions and open the folder that they are in.  It is designed to be used as a runnable JAR or applet that should be placed in the highest level folder for searching. Cannot be sure of how quickly the search will return in extremely large systems.